pyproject.toml
[tool.poetry]
name = "youtubedownloader"
version = "0.1.0"
description = ""
authors = ["LordGvozd"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
pafy = "^0.5.5"
youtube-dl = "^2021.12.17"
pytube = "^15.0.0"
fastapi = "^0.111.0"
fastui = "^0.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
mypy = "^1.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
